Move classes to another package to clean things
