+++ src/main/java/com/meti/app/GetCommandEvaluatable.java
@@ -1,6 +1,8 @@
 package com.meti.app;
 
 import com.meti.lib.net.command.GetCommand;
+import com.meti.lib.net.server.evaluate.AbstractEvaluatable;
+import com.meti.lib.net.server.evaluate.Evaluator;
 import com.meti.lib.util.StreamUtil;
 
 import java.io.Serializable;
