+++ src/main/java/com/meti/lib/net/server/evaluate/Evaluatable.java
@@ -1,4 +1,4 @@
-package com.meti.app;
+package com.meti.lib.net.server.evaluate;
 
 import java.io.Serializable;
