+++ src/main/java/com/meti/lib/net/server/evaluate/Evaluator.java
@@ -1,4 +1,4 @@
-package com.meti.app;
+package com.meti.lib.net.server.evaluate;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
